+++ src/frontend/index.less
@@ -15,8 +15,8 @@
 @import "./komponenter/Fagsak/Høyremeny/høyremeny.less";
 
 @import "./komponenter/Fagsak/Vilkårsvurdering/vilkårsvurdering.less";
-@import "./komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/generisk-vilkår.less";
-@import "./komponenter/Fagsak/Vilkårsvurdering/GeneriskVilkår/FastsettPeriode/fastsett-periode.less";
+@import "src/frontend/komponenter/Fagsak/Vilkårsvurdering/VilkårBolk/vilkår-tabell.less";
+@import "src/frontend/komponenter/Fagsak/Vilkårsvurdering/VilkårBolk/FastsettPeriode/fastsett-periode.less";
 @import "./komponenter/Fagsak/Totrinnskontroll/totrinnskontroll.less";
 
 // Felles
