+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/VilkårBolk/FastsettPeriode/fastsett-periode.less
