+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/VilkårBolk/VilkårTabellRad.tsx
@@ -1,0 +1,130 @@
+import React, { useState } from 'react';
+
+import deepEqual from 'deep-equal';
+
+import { Normaltekst } from 'nav-frontend-typografi';
+
+import { FeltState } from '@navikt/familie-skjema';
+import { RessursStatus } from '@navikt/familie-typer';
+
+import { useBehandling } from '../../../../context/BehandlingContext';
+import AutomatiskVurdering from '../../../../ikoner/AutomatiskVurdering';
+import FamilieChevron from '../../../../ikoner/FamilieChevron';
+import ManuellVurdering from '../../../../ikoner/ManuellVurdering';
+import VilkårResultatIkon from '../../../../ikoner/VilkårResultatIkon';
+import { periodeToString } from '../../../../typer/periode';
+import { IGrunnlagPerson } from '../../../../typer/person';
+import { IVilkårConfig, IVilkårResultat, Resultat, uiResultat } from '../../../../typer/vilkår';
+import { datoformat, formaterIsoDato } from '../../../../utils/formatter';
+import IkonKnapp from '../../../Felleskomponenter/IkonKnapp/IkonKnapp';
+import { vilkårFeilmeldingId } from './VilkårTabell';
+import VilkårTabellRadEndre from './VilkårTabellRadEndre';
+
+interface IProps {
+    person: IGrunnlagPerson;
+    vilkårFraConfig: IVilkårConfig;
+    vilkårResultat: FeltState<IVilkårResultat>;
+    visFeilmeldinger: boolean;
+}
+
+const VilkårTabellRad: React.FC<IProps> = ({
+    person,
+    vilkårFraConfig,
+    vilkårResultat,
+    visFeilmeldinger,
+}) => {
+    const { erLesevisning, åpenBehandling } = useBehandling();
+
+    const [ekspandertVilkår, settEkspandertVilkår] = useState(
+        erLesevisning() || false || vilkårResultat.verdi.resultat.verdi === Resultat.IKKE_VURDERT
+    );
+    const [redigerbartVilkår, settRedigerbartVilkår] = useState<FeltState<IVilkårResultat>>(
+        vilkårResultat
+    );
+
+    const toggleForm = (visAlert: boolean) => {
+        if (ekspandertVilkår && visAlert && !deepEqual(vilkårResultat, redigerbartVilkår)) {
+            alert('Vurderingen har endringer som ikke er lagret!');
+        } else {
+            settEkspandertVilkår(!ekspandertVilkår);
+            settRedigerbartVilkår(vilkårResultat);
+        }
+    };
+
+    return (
+        <tbody>
+            <tr className={ekspandertVilkår ? 'ekspandert' : ''}>
+                <td>
+                    <div className={'vurdering'}>
+                        <VilkårResultatIkon
+                            resultat={vilkårResultat.verdi.resultat.verdi}
+                            width={20}
+                            heigth={20}
+                        />
+                        <Normaltekst children={uiResultat[vilkårResultat.verdi.resultat.verdi]} />
+                    </div>
+                </td>
+                <td>
+                    <Normaltekst children={periodeToString(vilkårResultat.verdi.periode.verdi)} />
+                </td>
+                <td>
+                    <Normaltekst
+                        className={'beskrivelse'}
+                        children={vilkårResultat.verdi.begrunnelse.verdi}
+                    />
+                </td>
+                <td>
+                    <IkonKnapp
+                        erLesevisning={erLesevisning()}
+                        onClick={() => toggleForm(true)}
+                        id={vilkårFeilmeldingId(vilkårResultat.verdi)}
+                        label={
+                            !ekspandertVilkår
+                                ? vilkårResultat.verdi.resultat.verdi === Resultat.IKKE_VURDERT
+                                    ? 'Vurder'
+                                    : 'Endre'
+                                : 'Lukk'
+                        }
+                        mini={true}
+                        ikon={<FamilieChevron retning={ekspandertVilkår ? 'opp' : 'ned'} />}
+                    />
+                </td>
+                <td>
+                    {vilkårResultat.verdi.erAutomatiskVurdert ? (
+                        <AutomatiskVurdering />
+                    ) : (
+                        <ManuellVurdering />
+                    )}
+                </td>
+                <td>
+                    <i>
+                        {åpenBehandling.status === RessursStatus.SUKSESS &&
+                        vilkårResultat.verdi.erVurdert
+                            ? vilkårResultat.verdi.behandlingId === åpenBehandling.data.behandlingId
+                                ? 'Vurdert i denne behandlingen'
+                                : `Vurdert ${formaterIsoDato(
+                                      vilkårResultat.verdi.endretTidspunkt,
+                                      datoformat.DATO_FORKORTTET
+                                  )}`
+                            : ''}
+                    </i>
+                </td>
+            </tr>
+
+            {ekspandertVilkår && (
+                <tr>
+                    <td colSpan={6} className={'td-ekspandert'}>
+                        <VilkårTabellRadEndre
+                            person={person}
+                            vilkårFraConfig={vilkårFraConfig}
+                            vilkårResultat={vilkårResultat}
+                            visFeilmeldinger={visFeilmeldinger}
+                        />
+                    </td>
+                </tr>
+            )}
+        </tbody>
+    );
+};
+
+export default VilkårTabellRad;
