+++ src/frontend/komponenter/Fagsak/Vilkårsvurdering/VilkårBolk/vilkår-tabell.less
@@ -1,4 +1,4 @@
-.generisk-vilkår {
+.vilkår-tabell {
     margin-top: 1rem;
 
     .tabell {
@@ -46,7 +46,7 @@
         }
     }
 
-    & ~ .generisk-vilkår {
+    & ~ .vilkår-tabell {
         margin-top: 2.5rem;
     }
